Chore/Cleaned up engine
Removed steamworks and imgui

+++ Minigin/LoggedSoundSystem.h
@@ -11,7 +11,7 @@ namespace FH
 		LoggedSoundSystem(SoundBase& wrappedSystem);
 		~LoggedSoundSystem() = default;
 
-		virtual void Play(soundId id, int volume = 128) override;
+		virtual void Play(soundId id, float volume = 128) override;
 		//virtual void PlayLooped(soundId id, float volume) override;
 		virtual void TryStop(soundId id) override;
 		virtual void StopAll() override;
