+++ Minigin/SoundBase.h
@@ -8,8 +8,8 @@ namespace FH
 	{
 	public:
 		virtual ~SoundBase() = default;
-		virtual void Play(soundId id, int volume) = 0;
-		//virtual void PlayLooped(soundId id, int volume) = 0;
+		virtual void Play(soundId id, float volume) = 0;
+		//virtual void PlayLooped(soundId id, float volume) = 0;
 		virtual void TryStop(soundId id) = 0;
 		virtual void StopAll() = 0;
 	};
@@ -17,8 +17,8 @@ namespace FH
 	class NullSoundSystem : public SoundBase
 	{
 	public:
-		virtual void Play(soundId id, int volume) override { id; volume; }
-		//virtual void PlayLooped(soundId id, int volume) override {}
+		virtual void Play(soundId id, float volume) override { id; volume; }
+		//virtual void PlayLooped(soundId id, float volume) override {}
 		virtual void TryStop(soundId id) override { id; }
 		virtual void StopAll() override	{}
 	};
