+++ Minigin/SoundSystem.h
@@ -13,8 +13,8 @@ namespace FH
 		SoundSystem& operator=(const SoundSystem&) = delete;
 		SoundSystem& operator=(SoundSystem&&) = default;
 
-		virtual void Play(soundId id, int volume = 128) override;
-		//virtual void PlayLooped(soundId id, int volume) override;
+		virtual void Play(soundId id, float volume = 128) override;
+		//virtual void PlayLooped(soundId id, float volume) override;
 		virtual void TryStop(soundId id) override;
 		virtual void StopAll() override;
